Add decimal funcinality.

+++ Calculator/Program.cs
@@ -1,4 +1,4 @@
-﻿int num1 = 0; int num2 = 0;
+﻿float num1 = 0; float num2 = 0;
 
 Console.WriteLine("Console Calculator in C#\r");
 Console.WriteLine("------------------------\n");
